Move text measurement into a neutral utility module out of the overlays layer
